requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.0
matplotlib==3.9.0
seaborn==0.13.2
pytest==8.2.1
starlette==0.37.2
httpx==0.24.1
structlog>=22.1.0,<26.0.0